requirements.txt
appdirs==1.4.4
black==22.3.0
click==8.1.3
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
pygubu==0.21
pygubu-designer==0.26
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
opencv-python==4.4.0.46
